requirements.txt
Django==1.7.6
SQLAlchemy==0.9.9
Unipath==1.1
argparse==1.2.1
django-user-agents==0.3.0
djangorestframework==3.1.0
psycopg2==2.6
ua-parser==0.3.6
user-agents==0.3.2
wsgiref==0.1.2
amqp==1.4.6
anyjson==0.3.3
billiard==3.3.0.19
celery==3.1.17
django-redis==3.8.3
django-redis-sessions==0.4.0
django-websocket-redis==0.4.3
gevent==1.0.1
greenlet==0.4.5
kombu==3.0.24
pytz==2014.10
redis==2.10.3
six==1.9.0

